Weapons properly related, needs seed work
