+++ app/controllers/weapons_controller.rb
@@ -1,0 +1,2 @@
+class WeaponsController < ApplicationController
+end
