+++ app/helpers/weapons_helper.rb
@@ -1,0 +1,2 @@
+module WeaponsHelper
+end
