+++ app/models/attached_weapon_property.rb
@@ -1,0 +1,4 @@
+class AttachedWeaponProperty < ApplicationRecord
+  belongs_to :weapon
+  belongs_to :weapon_property
+end
