+++ app/models/damage_type.rb
@@ -1,0 +1,3 @@
+class DamageType < ApplicationRecord
+  has_many :weapons
+end
