+++ app/models/weapon_category.rb
@@ -1,0 +1,3 @@
+class WeaponCategory < ApplicationRecord
+  has_many :weapons
+end
