+++ app/models/weapon_property.rb
@@ -1,0 +1,4 @@
+# An ActiveRecord class for storing weapon properties (light, finesse,etc.)
+class WeaponProperty < ApplicationRecord
+  has_many :weapons
+end
