+++ config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
+  resources :weapons
   resources :skills
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
@@ -10,5 +11,5 @@ Rails.application.routes.draw do
     resources :armors, only: [:index, :create]
   end
   resources :armors, except: [:create, :new]
-
 end
+
