+++ db/migrate/20171026215208_create_weapon_properties.rb
@@ -1,0 +1,10 @@
+class CreateWeaponProperties < ActiveRecord::Migration[5.1]
+  def change
+    create_table :weapon_properties do |t|
+      t.string :name
+      t.text :description
+
+      t.timestamps
+    end
+  end
+end
