+++ db/migrate/20171027170510_add_weapon_property_id_to_weapon.rb
@@ -1,0 +1,5 @@
+class AddWeaponPropertyIdToWeapon < ActiveRecord::Migration[5.1]
+  def change
+    add_reference :weapons, :weapon_property, foreign_key: true
+  end
+end
