+++ db/migrate/20171027171055_remove_property_id_from_weapon.rb
@@ -1,0 +1,5 @@
+class RemovePropertyIdFromWeapon < ActiveRecord::Migration[5.1]
+  def change
+    remove_column :weapons, :property_id
+  end
+end
