+++ db/migrate/20171101213429_add_weapon_category_id_to_weapons.rb
@@ -1,0 +1,5 @@
+class AddWeaponCategoryIdToWeapons < ActiveRecord::Migration[5.1]
+  def change
+    add_reference :weapons, :weapon_category, foreign_key: true
+  end
+end
