+++ db/migrate/20171101213514_remove_weapon_categories_id_from_weapons.rb
@@ -1,0 +1,5 @@
+class RemoveWeaponCategoriesIdFromWeapons < ActiveRecord::Migration[5.1]
+  def change
+    remove_column :weapons, :weapon_categories_id
+  end
+end
