+++ db/migrate/20171101214849_drop_table_damage_types.rb
@@ -1,0 +1,5 @@
+class DropTableDamageTypes < ActiveRecord::Migration[5.1]
+  def change
+    drop_table :damagetypes
+  end
+end
