+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171025181954) do
+ActiveRecord::Schema.define(version: 20171102045720) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -52,6 +52,15 @@ ActiveRecord::Schema.define(version: 20171025181954) do
     t.index ["armor_category_id"], name: "index_armors_on_armor_category_id"
   end
 
+  create_table "attached_weapon_properties", force: :cascade do |t|
+    t.bigint "weapon_id"
+    t.bigint "weapon_property_id"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["weapon_id"], name: "index_attached_weapon_properties_on_weapon_id"
+    t.index ["weapon_property_id"], name: "index_attached_weapon_properties_on_weapon_property_id"
+  end
+
   create_table "character_classes", force: :cascade do |t|
     t.string "name"
     t.string "hit_die", limit: 3
@@ -60,6 +69,13 @@ ActiveRecord::Schema.define(version: 20171025181954) do
     t.string "hit_points_at_first"
     t.string "hit_points_at_higher"
     t.integer "skill_amount"
+  end
+
+  create_table "damage_types", force: :cascade do |t|
+    t.string "name"
+    t.string "description"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
   end
 
   create_table "save_proficiencies", force: :cascade do |t|
@@ -89,12 +105,42 @@ ActiveRecord::Schema.define(version: 20171025181954) do
     t.index ["ability_id"], name: "index_skills_on_ability_id"
   end
 
+  create_table "weapon_categories", force: :cascade do |t|
+    t.string "name"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  create_table "weapon_properties", force: :cascade do |t|
+    t.string "name"
+    t.text "description"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
+  create_table "weapons", force: :cascade do |t|
+    t.string "name"
+    t.string "cost"
+    t.string "damage_dice", limit: 3
+    t.integer "damage_type_id"
+    t.bigint "weapon_category_id"
+    t.string "weight"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["damage_type_id"], name: "index_weapons_on_damage_type_id"
+    t.index ["weapon_category_id"], name: "index_weapons_on_weapon_category_id"
+  end
+
   add_foreign_key "armor_proficiencies", "armor_categories"
   add_foreign_key "armor_proficiencies", "character_classes"
   add_foreign_key "armors", "armor_categories"
+  add_foreign_key "attached_weapon_properties", "weapon_properties"
+  add_foreign_key "attached_weapon_properties", "weapons"
   add_foreign_key "save_proficiencies", "abilities"
   add_foreign_key "save_proficiencies", "character_classes"
   add_foreign_key "skill_proficiencies", "character_classes"
   add_foreign_key "skill_proficiencies", "skills"
   add_foreign_key "skills", "abilities"
-end
+  add_foreign_key "weapons", "damage_types"
+  add_foreign_key "weapons", "weapon_categories"
+  end
