+++ test/fixtures/attached_weapon_properties.yml
@@ -1,0 +1,9 @@
+# Read about fixtures at http://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  weapon: one
+  weapon_property: one
+
+two:
+  weapon: two
+  weapon_property: two
