+++ test/fixtures/weapons.yml
@@ -1,0 +1,17 @@
+# Read about fixtures at http://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
+
+one:
+  name: MyString
+  cost: MyString
+  damage_dice: 
+  damage_type_id: 1
+  weight: MyString
+  property_id: 1
+
+two:
+  name: MyString
+  cost: MyString
+  damage_dice: 
+  damage_type_id: 1
+  weight: MyString
+  property_id: 1
