+++ test/models/attached_weapon_property_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class AttachedWeaponPropertyTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
