+++ test/models/damage_type_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class DamageTypeTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
