+++ test/models/weapon_category_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class WeaponCategoryTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
