+++ test/models/weapon_property_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class WeaponPropertyTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
