+++ test/models/weapon_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class WeaponTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
